+++ src/Exception/FeatureNotSupportedException.php
@@ -1,0 +1,13 @@
+<?php
+
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\EncoreBundle\Exception;
+
+class FeatureNotSupportedException extends \Exception
+{
+}